refactor: improve testability dividing page by logical parts
feat: implement EditPublicationsView
feat: implement usePublicationsManager hook
chore: connect usePublicationsManager hook and EditPublicationsView in
page.tsx
